Allows null values for setter parameters
Updates setter methods in Address, Company, FinalRecipient, Invoice, InvoiceItem, ItemVatSummary, NonStockItem, Partner and VatSummary classes to accept null values.

This change enhances the flexibility of the data model by allowing optional properties to be explicitly set to null, providing a more consistent way to represent the absence of a value.